feat: 로또 결과 계산 기능 정의
# step 3
