+++ src/test/kotlin/lotto/controller/EvaluateSpec.kt
@@ -2,22 +2,27 @@ package lotto.controller
 
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
+import lotto.domain.LottoNumber
+import lotto.domain.LottoTickets
+import lotto.presentation.controller.EvaluateRequest
+import lotto.presentation.controller.LottoController
 
-class EvaluateSpec: BehaviorSpec({
+class EvaluateSpec : BehaviorSpec({
     given("로또 구매가 완료되었을 때") {
         val lottoController = LottoController()
         lottoController.tickets = listOf(
-            LottoTicket(listOf(1, 2, 3, 4, 5, 6)),
-            LottoTicket(listOf(7, 8, 9, 10, 11, 12))
-        )
+            LottoNumber.of(listOf(1, 2, 3, 4, 5, 6)),
+            LottoNumber.of(listOf(7, 8, 9, 10, 11, 12))
+        ).let(::LottoTickets)
 
         `when`("우승자 번호와 보너스 볼로 평가 요청을 보냈을 때") {
             val evaluateRequest = EvaluateRequest.from("1, 2, 3, 4, 5, 6", "7")
 
             then("매칭 결과에 따라 Prize 획득 별로 집계된다.") {
-                val lottoResult = lottoController.evaluate(evaluateRequest)
-                lottoResult.rankCounts[Rank.FIRST] shouldBe 1
-                lottoResult.rankCounts[Rank.MISS] shouldBe 1
+                val evaluateResp = lottoController.evaluate(evaluateRequest)
+                evaluateResp.rankResult[0][0] shouldBe 6
+                evaluateResp.rankResult[0][1] shouldBe 2_000_000_000
+                evaluateResp.rankResult[0][2] shouldBe 1
             }
         }
     }
